requirements.txt
pip
numpy
streamlit
openai
sounddevice
scipy
wave
python-dotenv
requests